+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
   - New option page using options V2 standard
   - Support all kind of separators in separatethis.com
   - Remove option: `Bookmarklet supported`, now we always support bookmarklet WITHOUT scary permissions
+  - Remove permission: `Read and change all your data on the websites you visit`, now it replaced by `Read the icons of the websites you visit`!
 
 Bugfixes:
   - Cannot `Copy` a folder