+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Features:
   - Use `Tab` and `Shift + Tab` to navigate bookmarks
 
 Changes:
-  - Increase minimum Chrome version from 26 to 32
+  - Increase minimum Chrome version from 26 to 34
   - New option page using options V2 standard
   - Support all kind of separators in separatethis.com
   - Remove option: `Bookmarklet supported`, now we always support bookmarklet WITHOUT scary permissions